read_epw is now in a separate module instead of part of tmy.py
includes a couple of other small fixes and updated testing files
