+++ pvlib/iotools/__init__.py
@@ -1,6 +1,6 @@
 from pvlib.iotools.tmy import read_tmy2  # noqa: F401
 from pvlib.iotools.tmy import read_tmy3  # noqa: F401
-from pvlib.iotools.tmy import read_epw  # noqa: F401
+from pvlib.iotools.epw import read_epw  # noqa: F401
 from pvlib.iotools.srml import read_srml  # noqa: F401
 from pvlib.iotools.srml import read_srml_month_from_solardat  # noqa: F401
 from pvlib.iotools.surfrad import read_surfrad  # noqa: F401
